+++ CHANGELOG.md
@@ -4,13 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## 2.7.0 - Unreleased
+## 2.7.0 - 2018-08-19
+### Added
+- `add-network-policy` command - Create policy to allow direct network traffic from one app to another
+- `remove-network-policy` command - Remove network traffic policy of an app
+
 ### Changed
+- `create-user-provided-service` command now updates the ups if it already exists
 - Updated to [cf cli v6.37.0](https://github.com/cloudfoundry/cli/releases/tag/v6.37.0)
 - Updated to [autopilot cf plugin v0.0.6](https://github.com/contraband/autopilot/releases/tag/0.0.6)
 - Updated to [yq cli v2.1.0](https://github.com/mikefarah/yq/releases/tag/2.1.0)
-- `create-user-provided-service` now updates the ups if it already exists
-
 
 ## 2.6.0 - 2018-05-15
 ### Added